Separate settings from their UI
