+++ src/js/views/draw-chromosomes.js
@@ -1,5 +1,5 @@
 import {d3} from '../lib';
-import {initSettings} from '../init/settings';
+import {initSettings} from '../init/settings-ui';
 import {ModelAdapter} from '../model-adapter';
 import {Chromosome} from './chromosome';
